Added basic input handler.
Since a lot of the program will be about how well we can handle the user's
inputs, this has been abstracted away into a separate classed called
"InputHandler".

This will contain all functionality to handle inputs, as well as the
provided "readEntry" method. However, this method has been made private
for the sake of forcing proper handling of each case.